+++ app/CHANGELOG.md
@@ -1,3 +1,9 @@
+## What's Changed in v1.11.5
+
+- FIX: WebRTC - signaling URL to match request host. #684 Thanks @JA16122000!
+- FIX: WebUI - don't start socket connection until play/don't autoplay if on-demand cam is NOT connected. #690
+- NEW: WebUI - add toggle to switch between HLS/WebRTC.
+
 ## What's Changed in v1.11.4
 
 - FIX: Switching between webrtc/hls was broken in the WebUI (`?webrtc`/`?hls`)